+++ js/main.js
@@ -3,7 +3,7 @@ import { setCloseBigPhoto, setOpenBigPhoto } from './big-photo.js';
 import { setOverlayForm, validateForm, setUserFormSubmit } from './form.js';
 import { setScaleSizePhoto, setScaleEffectsPhoto } from './set-effects-photo.js';
 import { showAlert } from './message.js';
-import { getData } from './get-data.js';
+import { getData } from './api.js';
 getData(renderPhotos, setOpenBigPhoto, setCloseBigPhoto, showAlert);
 validateForm();
 setOverlayForm();
